Fixed the dropping items issues.
I accidentally had an assignment = instead of an equality == in one of the
conditionals, which was responsible for the big errors. The other small
thing was that because I didn't wait for a keypress, the player warning
dialogue got cleared. Right now I am not sure I want to wait for keypress,
in which case I will have to fiddle around with the order of clearing the
screen.